requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
openai>=2.20.0
langgraph>=0.1.0
langchain-openai>=1.1.0
python-dotenv>=1.0.0
